+++ portland-or-usa/your-name.md
@@ -1,0 +1,42 @@
+# Your Name Here
+
+![an avatar, if you'd like to include one](https://1.gravatar.com/avatar/8310c4341b623f781babb080cd2c5d4f?r=x&s=200)
+
+| **location**: | city, state, country |
+|:--------------|---------------------:|
+| **twitter**:  | @therealgarybusey    |
+| **github**:   | your-gh-username     |
+| **urls**:     | urls                 |
+| **travel availability**: | if comp'd / regionally / locally / none |
+| **preferred audience size**: | large / small / N/A |
+
+## About me
+
+A brief introduction. What are you into? What do you like to talk about / work on?
+Why do you enjoy talking? Are there any particular formats you enjoy?
+
+## Talks / Workshops (optional)
+
+### Something about JavaScript
+
+| **location**: | event, city, state, country |
+|:--------------|----------------------------:|
+| **when**:     | YYYY-MM-DD                  |
+| **materials**:   | link to speakerdeck, etc |
+| **duration**:   | 45-min / 30 min |
+| **audience size**: | large / small / ??? |
+
+A brief synopsis of what you talked about. E.G., I talked about XYZ at ABC. 
+The audience seemed to really like it. There were a few questions and they
+lead me to add some feature to the thing.
+
+### Bleep Bloop
+
+| **location**: | event, city, state, country |
+|:--------------|----------------------------:|
+| **when**:     | YYYY-MM-DD                  |
+| **materials**:   | link to speakerdeck, etc |
+| **duration**:   | 3 months |
+| **audience size**: | large / small / ??? |
+
+I taught robots to make squid noises using [workshopper](http://npm.im/workshopper).
